+++ spec/rubocop/cop/rspec/contain_exactly_spec.rb
@@ -45,17 +45,11 @@ RSpec.describe RuboCop::Cop::RSpec::ContainExactly do
     RUBY
   end
 
-  it 'flags `contain_exactly` with no arguments' do
-    expect_offense(<<-RUBY)
+  # Don't do the work of RSpec/BeEmpty
+  it 'does not flag `contain_exactly` with no arguments' do
+    expect_no_offenses(<<-RUBY)
       it { is_expected.to contain_exactly }
-                          ^^^^^^^^^^^^^^^ Prefer `be_empty` when matching an empty collection.
       it { is_expected.to contain_exactly() }
-                          ^^^^^^^^^^^^^^^^^ Prefer `be_empty` when matching an empty collection.
-    RUBY
-
-    expect_correction(<<-RUBY)
-      it { is_expected.to be_empty }
-      it { is_expected.to be_empty }
     RUBY
   end
 end